Refactor: Extract create_table.rs tests to separate file
- Move all 10 test functions from create_table.rs to tests/create_table_tests.rs
- Reduce create_table.rs from 541 to 140 lines (74% reduction)
- Improve code organization following Rust best practices
- All tests pass with no behavior changes

This improves code readability by separating implementation (~140 lines)
from tests (~421 lines), making the implementation easier to review and navigate.
